+++ src/Sentry/Laravel/Integration.php
@@ -30,7 +30,7 @@ class Integration implements IntegrationInterface
      */
     public function setupOnce(): void
     {
-        Scope::addGlobalEventProcessor(function (Event $event): Event {
+        Scope::addGlobalEventProcessor(static function (Event $event): Event {
             $self = SentrySdk::getCurrentHub()->getIntegration(self::class);
 
             if (!$self instanceof self) {
@@ -38,7 +38,7 @@ class Integration implements IntegrationInterface
             }
 
             if (empty($event->getTransaction())) {
-                $event->setTransaction($self->getTransaction());
+                $event->setTransaction(self::getTransaction());
             }
 
             return $event;
@@ -121,50 +121,116 @@ class Integration implements IntegrationInterface
      *
      * @param \Illuminate\Routing\Route $route
      *
-     * @return string|null
+     * @return string
      */
-    public static function extractNameForRoute(Route $route): ?string
+    public static function extractNameForRoute(Route $route): string
     {
         $routeName = null;
 
-        if (empty($routeName) && $route->getName()) {
-            // someaction (route name/alias)
-            $routeName = $route->getName();
-
-            // Laravel 7 route caching generates a route names if the user didn't specify one
-            // theirselfs to optimize route matching. These route names are useless to the
-            // developer so if we encounter a generated route name we discard the value
-            if (Str::contains($routeName, 'generated::')) {
-                $routeName = null;
-            }
-
-            // If the route name ends with a `.` we assume an incomplete group name prefix
-            // we discard this value since it will most likely not mean anything to the
-            // developer and will be duplicated by other unnamed routes in the group
-            if (Str::endsWith($routeName, '.')) {
-                $routeName = null;
-            }
+        // someaction (route name/alias)
+        if ($route->getName()) {
+            $routeName = self::extractNameForNamedRoute($route->getName());
         }
 
+        // Some\Controller@someAction (controller action)
         if (empty($routeName) && $route->getActionName()) {
-            // Some\Controller@someAction (controller action)
-            $routeName = ltrim($route->getActionName(), '\\');
-
-            $baseNamespace = self::$baseControllerNamespace ?? '';
-
-            // Strip away the base namespace from the action name
-            if (!empty($baseNamespace)) {
-                // @see: Str::after, but this is not available before Laravel 5.4 so we use a inlined version
-                $routeName = array_reverse(explode($baseNamespace . '\\', $routeName, 2))[0];
-            }
+            $routeName = self::extractNameForActionRoute($route->getActionName());
         }
 
+        // /someaction // Fallback to the url
         if (empty($routeName) || $routeName === 'Closure') {
-            // /someaction // Fallback to the url
             $routeName = '/' . ltrim($route->uri(), '/');
         }
 
         return $routeName;
+    }
+
+    /**
+     * Extract the readable name for a Lumen route.
+     *
+     * @param array  $routeData The array of route data
+     * @param string $path      The path of the request
+     *
+     * @return string
+     */
+    public static function extractNameForLumenRoute(array $routeData, string $path): string
+    {
+        $routeName = null;
+
+        $route = $routeData[1] ?? [];
+
+        // someaction (route name/alias)
+        if (!empty($route['as'])) {
+            $routeName = self::extractNameForNamedRoute($route['as']);
+        }
+
+        // Some\Controller@someAction (controller action)
+        if (empty($routeName) && !empty($route['uses'])) {
+            $routeName = self::extractNameForActionRoute($route['uses']);
+        }
+
+        // /someaction // Fallback to the url
+        if (empty($routeName) || $routeName === 'Closure') {
+            $routeUri = array_reduce(
+                array_keys($routeData[2]),
+                static function ($carry, $key) use ($routeData) {
+                    return str_replace($routeData[2][$key], "{{$key}}", $carry);
+                },
+                $path
+            );
+
+            $routeName = '/' . ltrim($routeUri, '/');
+        }
+
+        return $routeName;
+    }
+
+    /**
+     * Take a route name and return it only if it's a usable route name.
+     *
+     * @param string $name
+     *
+     * @return string|null
+     */
+    private static function extractNameForNamedRoute(string $name): ?string
+    {
+        // Laravel 7 route caching generates a route names if the user didn't specify one
+        // theirselfs to optimize route matching. These route names are useless to the
+        // developer so if we encounter a generated route name we discard the value
+        if (Str::contains($name, 'generated::')) {
+            return null;
+        }
+
+        // If the route name ends with a `.` we assume an incomplete group name prefix
+        // we discard this value since it will most likely not mean anything to the
+        // developer and will be duplicated by other unnamed routes in the group
+        if (Str::endsWith($name, '.')) {
+            return null;
+        }
+
+        return $name;
+    }
+
+    /**
+     * Take a controller action and strip away the base namespace if needed.
+     *
+     * @param string $action
+     *
+     * @return string
+     */
+    private static function extractNameForActionRoute(string $action): string
+    {
+        $routeName = ltrim($action, '\\');
+
+        $baseNamespace = self::$baseControllerNamespace ?? '';
+
+        if (empty($baseNamespace)) {
+            return $routeName;
+        }
+
+        // Strip away the base namespace from the action name
+        // @see: Str::after, but this is not available before Laravel 5.4 so we use a inlined version
+        return array_reverse(explode($baseNamespace . '\\', $routeName, 2))[0];
     }
 
     /**