+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- Add support for normalized route names when using Lumen (#449)
+
+## 2.11.1
+
+- Fix deprecation notice in route name extraction (#543)
+
 ## 2.11.0
 
 - Add support for Laravel 9 (#534)